# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Flood Warning System",
    "version": "3.0.0"
  },
  "paths": {
...
    "/response/events": {
      "get": {
        "tags": [
          "district-response-workflow"
        ],
        "summary": "List Events",
        "operationId": "list_events_response_events_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "district-response-workflow"
        ],
        "summary": "Create Event",
        "operationId": "create_event_response_events_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/flood_system__response_workflow__models__EventCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/response/events/{event_id}": {
      "get": {
        "tags": [
          "district-response-workflow"
        ],
        "summary": "Get Event Dashboard",
        "operationId": "get_event_dashboard_response_events__event_id__get",
        "parameters": [
          {
            "name": "event_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Event Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/response/events/{event_id}/integrity": {
      "get": {
        "tags": [
          "district-response-workflow"
        ],
        "summary": "Verify Event Integrity",
        "operationId": "verify_event_integrity_response_events__event_id__integrity_get",
        "parameters": [
          {
            "name": "event_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Event Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/response/events/{event_id}/alerts": {
      "post": {
        "tags": [
          "district-response-workflow"
        ],
        "summary": "Append Alert",
        "operationId": "append_alert_response_events__event_id__alerts_post",
        "parameters": [
          {
            "name": "event_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Event Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AlertAppendRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/response/events/{event_id}/risk-objects": {
      "post": {
        "tags": [
          "district-response-workflow"
        ],
        "summary": "Add Risk Objects",
        "operationId": "add_risk_objects_response_events__event_id__risk_objects_post",
        "parameters": [
          {
            "name": "event_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Event Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RiskObjectBatchRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/response/events/{event_id}/risk-objects/discover": {
      "post": {
        "tags": [
          "district-response-workflow"
        ],
        "summary": "Discover Risk Objects",
        "operationId": "discover_risk_objects_response_events__event_id__risk_objects_discover_post",
        "parameters": [
          {
            "name": "event_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Event Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CandidateDiscoveryRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/response/events/{event_id}/candidate-runs": {
      "get": {
        "tags": [
          "district-response-workflow"
        ],
        "summary": "List Candidate Runs",
        "operationId": "list_candidate_runs_response_events__event_id__candidate_runs_get",
        "parameters": [
          {
            "name": "event_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Event Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/response/events/{event_id}/evidence-packages": {
      "get": {
        "tags": [
          "district-response-workflow"
        ],
        "summary": "List Evidence Packages",
        "operationId": "list_evidence_packages_response_events__event_id__evidence_packages_get",
        "parameters": [
          {
            "name": "event_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Event Id"
            }
          },
          {
            "name": "object_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Object Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/response/evidence-packages/{package_id}": {
      "get": {
        "tags": [
          "district-response-workflow"
        ],
        "summary": "Get Evidence Package",
        "operationId": "get_evidence_package_response_evidence_packages__package_id__get",
        "parameters": [
          {
            "name": "package_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Package Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "OperatorRole": {
        "type": "string",
        "enum": [
          "duty_officer",
          "reviewer",
          "commander",
          "liaison",
          "field_operator",
          "auditor",
          "admin",
          "external_service"
        ],
        "title": "OperatorRole"
      },
...
      "flood_system__response_workflow__models__EventCreateRequest": {
        "properties": {
          "title": {
            "type": "string",
            "title": "Title"
          },
          "area_id": {
            "type": "string",
            "title": "Area Id"
          },
          "alert": {
            "$ref": "#/components/schemas/AlertInput"
          },
          "operator_id": {
            "type": "string",
            "title": "Operator Id"
          },
          "operator_role": {
            "$ref": "#/components/schemas/OperatorRole"
          },
          "terminal_id": {
            "type": "string",
            "title": "Terminal Id",
            "default": "unknown-terminal"
          }
        },
        "type": "object",
        "required": [
          "title",
          "area_id",
          "alert",
          "operator_id",
          "operator_role"
        ],
        "title": "EventCreateRequest"
      },
...
    }
  }
}